+++ R/helpers.R
@@ -26,8 +26,8 @@ validate_colors <- function(opts) {
 }
 
 ## should make this return something that will be evaluated at render time
-get_next_color <- function(lgroupobj, which = "fill_color", type = "discrete") {
-  cur_theme <- bk_theme[[which]][[type]](10)
+get_next_color <- function(lgroupobj, which = "fill_color", type = "discrete", theme) {
+  cur_theme <- theme[[type]][[which]](10)
   n_layers <- length(lgroupobj$glyph_ids) + 1
   next_color_idx <- (n_layers - 1) %% length(cur_theme) + 1
   cur_theme[next_color_idx]
@@ -267,7 +267,7 @@ resolve_line_args <- function(fig, args) {
     args$line_cap <- ljoin_dict[[as.character(args$line_cap)]]
 
   if(is.null(args$line_color))
-    args$line_color <- get_next_color(fig)
+    args$line_color <- get_next_color(fig, theme = fig$x$spec$theme)
 
   args$color <- NULL
   args$alpha <- NULL
@@ -276,11 +276,11 @@ resolve_line_args <- function(fig, args) {
 }
 
 ## take args color and alpha and translate them to f
-resolve_color_alpha <- function(args, has_line = TRUE, has_fill = TRUE, ly, solid = FALSE) {
+resolve_color_alpha <- function(args, has_line = TRUE, has_fill = TRUE, ly, solid = FALSE, theme = NULL) {
 
   ## if no color at all is specified, choose from the theme
   if(is.null(args$color) && is.null(args$fill_color) && is.null(args$line_color))
-    args$color <- get_next_color(ly)
+    args$color <- get_next_color(ly, theme = theme)
 
   if(!is.null(args$color)) {
     if(!is.null(args$line_color)) {